requirements.txt
# -------------------------
# Platform notes
# -------------------------
# Python: 3.10 recommended
# Windows: Conda recommended for binary stability
# Pin NumPy < 2.0 to avoid SciPy/sklearn binary breakage.

# -------------------------
# Core API
# -------------------------
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27,<0.31
pydantic>=2.5,<3

# -------------------------
# UI
# -------------------------
streamlit>=1.33,<2
requests>=2.31,<3
pandas>=2.1,<2.3
matplotlib>=3.8,<3.10

# -------------------------
# Scientific stack (pins matter on Windows)
# -------------------------
numpy>=1.23,<2.0
scipy>=1.10,<1.12
scikit-learn>=1.3,<1.5

# -------------------------
# HF / NLP
# -------------------------
datasets>=2.18,<3
transformers>=4.40,<5
accelerate>=0.28,<1
sentencepiece>=0.1.99
sentence-transformers>=2.6,<3

# -------------------------
# Vector search
# -------------------------
faiss-cpu>=1.7.4,<2

# -------------------------
# PyTorch (CPU default)
# -------------------------
torch>=2.1,<2.4
torchvision>=0.16,<0.19
torchaudio>=2.1,<2.4